close files on certain OS when eof
Linux and possibly other now-windows OS do not emit a remove event when a file is deleted until the last file descriptor to that file is closed (because it watches an inode not a path).

So we close and try and reopen the file if we hit EOF on those systems as a bandaid for quickly checking 'does this file still exist'.